+++ src/modules/altran/elements/GnomeList/GnomeList.test.tsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { render } from '@testing-library/react'
+
+import GnomeList from './GnomeList';
+
+
+test('It Renders without crashing', () => {
+  const { container } = render(<GnomeList data={[]} loading={false} />)
+  expect(container.firstChild).toBeInTheDocument(); 
+})